requirements.txt
numpy
scipy
matplotlib
pywavelets
torch
torchvision
scikit-learn
pyyaml
tqdm
